Day2/ecommerce_app/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
python-dotenv
pydantic
psycopg2-binary